+++ src/Platform/Dashboard.php
@@ -17,7 +17,7 @@ class Dashboard
     /**
      * ORCHID Version.
      */
-    public const VERSION = '7.14.1';
+    public const VERSION = '7.15.0';
 
     /**
      * The Dashboard configuration options.