+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 7.15.0 - 2020-07-30
+
+### Added
+- Added `Chartable` trait
+
+### Deprecated
+- Property `labels` in `Chart`
+
 ## 7.14.1 - 2020-07-27
 
 ### Fixed